+++ drizzle.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
-  schema: './src/main/database/drizzle/schema',
+  schema: './src/main/database/drizzle/schema/*.ts',
   out: './drizzle',
   dialect: 'sqlite',
   dbCredentials: {